Structuring the files
